some shocking experimental result happened.
1. new[] and free(); were much faster than we thought.
- simple test on this test-suite involves repeated allocation and
freeing with two method - (MemoryPool / native method)

It reveals that new[] and free() are extremely fast
so that we barely don't need to use other methods
to alloc() and free().

However, maybe we can see difference when we make a
deliberately-generated set that induces thrashing.

That would be treated later.